fix windows compilation warnings
git-svn-id: http://svn.cern.ch/guest/lhcb/Rec/trunk@87481 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: 0678098bae9a3eb707ec4d5039d987e0f4b2cc67

+++ Rec/Calo/CaloMoniDst/src/CaloHypoNtp.cpp
@@ -153,7 +153,7 @@ StatusCode CaloHypoNtp::execute(){
       if( m_checker ){
         const LHCb::MCParticle* mcp = calo2MC()->from(hypo)->bestMC();
         if( NULL == mcp )continue;
-        if( m_mcID >= 0 && mcp->particleID().abspid()  != m_mcID)continue;        
+        if( m_mcID >= 0 && (int) mcp->particleID().abspid()  != m_mcID)continue;        
       }      
 
       // PrintOut
